backend/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
neo4j>=5.15.0

# Validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment
python-dotenv>=1.0.0

# HTTP client
httpx>=0.26.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.1.0

# SEC EDGAR
sec-edgar-downloader>=5.0.0

# Anthropic Claude API
anthropic>=0.18.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
ruff>=0.1.14
mypy>=1.8.0
